finish up filter styles
