+++ client/src/components/FilterSearch/index.js
@@ -40,15 +40,15 @@ function FilterSearch() {
                   <TypeOfPlace />
                 </div>
                 {/*footer*/}
-                <div className='flex items-center justify-end p-6 border-t border-solid border-slate-200 rounded-b'>
+                <div className='flex items-center justify-between p-6 border-t border-solid border-slate-200 rounded-b'>
                   <button
-                    className='text-red-500 background-transparent font-bold uppercase px-6 py-2 text-sm outline-none focus:outline-none mr-1 mb-1 ease-linear transition-all duration-150'
+                    className='underline text-black-500 background-transparent font-semibold  px-6 py-2 text-md outline-none focus:outline-none mr-1 mb-1 ease-linear transition-all duration-150'
                     type='button'
                     onClick={() => setShowModal(false)}>
-                    Close
+                    Clear All
                   </button>
                   <button
-                    className='bg-offBlack bg-emerald-500 text-white active:bg-emerald-600 font-bold uppercase text-sm px-6 py-3 rounded shadow hover:shadow-lg outline-none focus:outline-none mr-1 mb-1 ease-linear transition-all duration-150'
+                    className='bg-offBlack bg-emerald-500 text-white active:bg-emerald-600 font-bold text-md px-6 py-3 rounded shadow hover:shadow-lg outline-none focus:outline-none mr-1 mb-1 ease-linear transition-all duration-150'
                     type='button'
                     onClick={() => setShowModal(false)}>
                     Show Stays
